requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
bcrypt==3.2.0
cffi==1.15.0
cryptography==36.0.1
fastapi==0.73.0
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
iso8601==0.1.16
passlib==1.7.4
priority==2.0.0
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
pypika-tortoise==0.1.3
python-multipart==0.0.5
pytz==2021.3
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
toml==0.10.2
tortoise-orm==0.18.1
typing_extensions==4.1.1
wsproto==1.0.0
